+++ src/components/common/transaction-asset-icon/transaction-asset-icon.types.ts
@@ -1,0 +1,4 @@
+export enum IconSizeEnumType {
+  small = 'small',
+  large = 'large',
+}