+++ src/components/functional/toasts-list/components/transaction-toast/components/transaction-toast-content/transaction-toast-content.tsx
@@ -1,6 +1,8 @@
 import { faTimes } from '@fortawesome/free-solid-svg-icons';
 import type { EventEmitter, JSX } from '@stencil/core';
 import { Component, Event, h, Prop } from '@stencil/core';
+import { IconSizeEnumType } from 'components/common/transaction-asset-icon/transaction-asset-icon.types';
+import { getAmountParts } from 'components/functional/toasts-list/helpers';
 import type { ITransactionListItem } from 'components/visual/transaction-list-item/transaction-list-item.types';
 import { DataTestIdsEnum } from 'constants/dataTestIds.enum';
 
@@ -20,55 +22,13 @@ export class TransactionToastContent {
     this.deleteToast.emit();
   }
 
-  private renderPrimaryIcon() {
-    const transaction = this.transactions?.[0];
-
-    if (!transaction?.asset) {
-      return <mvx-default-transaction-icon-small />;
-    }
-
-    if (transaction.asset.imageUrl) {
-      return <img src={transaction.asset.imageUrl} alt="Transaction icon" class="transaction-toast-icon" loading="lazy" />;
-    }
-
-    if (transaction.asset.icon) {
-      return <mvx-fa-icon icon={transaction.asset.icon} class="transaction-toast-icon" />;
-    }
-
-    if (transaction.asset.text) {
-      return <span class="transaction-toast-icon">{transaction.asset.text}</span>;
-    }
-
-    if (this.toastDataState.icon) {
-      return <fa-icon icon={this.toastDataState.icon} class={`transaction-toast-icon ${this.toastDataState.iconClassName ?? ''}`}></fa-icon>;
-    }
-
-    return <mvx-default-transaction-icon-small />;
-  }
-
-  private renderDetails() {
-    const transaction = this.transactions?.[0];
-
-    if (!transaction) {
-      return null;
-    }
-
-    return (
-      <div class="transaction-toast-details-info">
-        {transaction.directionLabel && <span class="transaction-toast-details-info-text">{transaction.directionLabel}</span>}
-        <div class="transaction-toast-details-info-icon">
-          {transaction.interactorAsset ? <img src={transaction.interactorAsset} alt="Service icon" loading="lazy" /> : <mvx-default-transaction-icon-small />}
-        </div>
-        <mvx-trim-text text={transaction.interactor} class="transaction-toast-details-info-text" />
-      </div>
-    );
-  }
-
   render() {
     const { title, hasCloseButton } = this.toastDataState;
-    const transaction = this.transactions[0];
+    const [transaction] = this.transactions;
     const showAmount = this.transactions.length === 1 && transaction?.amount;
     const showExplorerLinkButton = transaction?.link && this.transactions.length === 1;
+    const amount = transaction && getAmountParts(transaction.amount);
+    const showPrimaryIcon = transaction.asset == null || transaction.asset.imageUrl || transaction.asset.icon || transaction.asset.text;
 
     return (
       <div
@@ -79,8 +39,13 @@ export class TransactionToastContent {
         data-testid={DataTestIdsEnum.transactionToastContent}
       >
         <div class="transaction-toast-content">
-          <div class="transaction-toast-icon">{this.renderPrimaryIcon()}</div>
-
+          {!showPrimaryIcon && this.toastDataState.icon ? (
+            <fa-icon icon={this.toastDataState.icon} class={`transaction-toast-icon ${this.toastDataState.iconClassName ?? ''}`}></fa-icon>
+          ) : (
+            <div class="transaction-toast-icon">
+              <mvx-transaction-asset-icon transaction={transaction} iconSize={IconSizeEnumType.small} />
+            </div>
+          )}
           <div class="transaction-toast-details">
             <div class="transaction-toast-details-header">
               <h4
@@ -89,11 +54,29 @@ export class TransactionToastContent {
                   'transaction-toast-title-short': Boolean(showAmount),
                 }}
               >
-                {title || transaction?.action?.name}
+                {transaction?.action?.name || title}
               </h4>
-              {showAmount && <div class="transaction-toast-amount">{transaction.amount}</div>}
+              {showAmount && (
+                <mvx-format-amount
+                  class="transaction-toast-amount"
+                  isValid
+                  label={amount.label}
+                  valueDecimal={amount.amountDecimal}
+                  valueInteger={amount.amountInteger}
+                  labelClass="transaction-amount-label"
+                  decimalClass="transaction-amount-decimal"
+                />
+              )}
             </div>
-            {this.renderDetails()}
+            {transaction && (
+              <div class="transaction-toast-details-info">
+                {transaction.directionLabel && <span class="transaction-toast-details-info-text">{transaction.directionLabel}</span>}
+                <div class="transaction-toast-details-info-icon">
+                  {transaction.interactorAsset ? <img src={transaction.interactorAsset} alt="Service icon" loading="lazy" /> : <mvx-default-transaction-icon-small />}
+                </div>
+                <mvx-trim-text text={transaction.interactor} class="transaction-toast-details-info-text" />
+              </div>
+            )}
           </div>
 
           {hasCloseButton && <fa-icon icon={faTimes} class="transaction-toast-close-icon" onClick={this.handleDeleteToast.bind(this)}></fa-icon>}